Weird MPI compiler bug.
When beta in block_inner isn't zero at compile time, the run hangs after
two iterations (on qe_erfc?) and then the stack blows up.  This only
occurs when the run is initiated with mpirun (even single-core).
Compiling with -O0 removes the issue.

This commit is to record the current state and facilitate testing on a
few other platforms.