Implementation of s_matrix and h_matrix with srb-matrix.
Projectors are distributed atom-wise across the cores in a pot. This
way, the RHSs Q \beta and D \beta are local and the LHSs
\beta (Q \beta) and \beta (D \beta) require a single reduce similar
to the inner product case in build_basis and h_coeff.  However,
The GEMMs now have a 'N', 'C' profile, so I'm calling it an outer
product and might have to do some optimizations later.

This outer product, along with a straight-forward add_diag, are
used to implement h_matrix and s_matrix.  It has been tested in
the single core case, but isnt' ready for multi-core.  This commit
is more of an exercise in design...