Making 3 blacs contexts; giving projectors a blacs desc.
The operation of computing the projections:
  <\beta|u> = \sum_b <\beta|b><b|u>
is difficult because both the projectors and the states are
distributed in their second dimension: by bands and projectors
respectively.  Thus, this operation requires redistribution.
However, since the data layout is the same, we can use the same
BLACS context and thus PBLAS.  This commit is mostly setting up
those contexts.

The projectors are distributed atom-wise: the types are separated
and the block size of each is set to nh(t).