Assorted fixes to allow npot /= nproc_pool until forces/stress.
The trickiest thing is making sure that the projectors, which are
distributed over atoms by type, handle 0 extents properly.  For
example, if there is a single atom single type distributed over
2 processors, projs(1)%dat will either be (nbasis, 0), or (0,basis)
on the second proc.  The former is what we need for the loops
to be skipped, which requires a (1,2) process grid.  If asymmetric,
npcol > nprow is preferable.

Force and stress hang.