Fixed stress for arbitrary proc/pot combinations.
The ugly thing about stress is that the terms rely on evc, which
is distributed pool-wise.  In the current code, that means globally.
Therefore, the projectors also need to be distributed globally.

Furthermore, stres_us expects a final mp_sum(intra_pool_comm), so
those sums can't be handled explicitly in the body of the code.

This works by redistributing states and bstates serially (globally),
transforming back out to PWs, and then using the normal codepath.