Print k-points, not q-points at start of file.  This addresses #6.
Note that the k-points printed in this way are only identical to those
coming from a conventional PW calculation up to a translation.  This
is because the SRB prefers to work in the unit cube BZ, not the
Wigner-Seitz cell.

The output eigenvalues still need to be printed for the k-points,
which this doesn't address.